django_base_project/requirements.txt
# Define all packages that are allowed to be installed externally
--allow-external mysql-connector-python

# Base 
setuptools
Django>=1.7
djangorestframework
django_compressor

# AWS 
boto
awsebcli

# Authentication
python-social-auth

# Async Tasks
celery
django-celery

# Testing
pytest
pytest-cov
pytest-django
mock
selenium

# Caching
django-redis

# Database
mysql-connector-python

# Misc
requests
